Renamed test_gcc_plugin to test_gcc_plugin_to_grfn since those tests run all the way to GrFN. Started new test structure to test the output of the gcc plugin, similar to the structure for test_gcc_ast_to_cast
